requirements.txt
fastapi>=0.111.0
krutrim-cloud==0.1.0
pydantic>=2.0.0
python-dotenv
python-docx
uvicorn>=0.30.0
langchain>=0.2.0
langchain-core>=0.2.0
langgraph>=0.1.0
